Added removed an event in database

+++ Calendar/src/ch/epfl/calendar/persistence/DBQuesterMock.java
@@ -63,7 +63,7 @@ public class DBQuesterMock implements LocalDatabaseInterface {
     }
 
 	@Override
-	public Event getEvent(int id) {
+	public Event getEvent(long id) {
 		return null;
 	}
